# src/problem5/openapi.yaml
openapi: 3.0.0
info:
  title: E-Commerce API
  version: 1.0.0
  description: A RESTful API for e-commerce operations - managing users, products, and orders
servers:
  - url: http://localhost:3000
    description: Development server
...
paths:
...
  /users/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          minimum: 1
        description: User ID

    get:
      summary: Get a user by ID
      description: Returns a single user by their ID
      tags:
        - Users
      responses:
        '200':
          description: User details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '404':
          description: User not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '429':
          description: Rate limit exceeded
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    patch:
      summary: Update a user
      description: Updates only the provided fields of an existing user
      tags:
        - Users
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateUser'
      responses:
        '200':
          description: User updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: User not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '409':
          description: Email already exists
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '429':
          description: Rate limit exceeded
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    delete:
      summary: Delete a user
      description: Permanently deletes a user
      tags:
        - Users
      responses:
        '204':
          description: User deleted successfully
        '404':
          description: User not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '429':
          description: Rate limit exceeded
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  # Products endpoints
...
components:
  schemas:
...
    User:
      type: object
      properties:
        id:
          type: integer
          description: Unique identifier
          example: 1
        email:
          type: string
          format: email
          description: User email address (unique)
          example: "john@example.com"
        name:
          type: string
          description: User's full name
          example: "John Doe"
        role:
          type: string
          enum: [customer, admin]
          description: User role
          example: "customer"
        createdAt:
          type: string
          format: date-time
          description: Creation timestamp (ISO 8601)
          example: "2024-01-15T10:30:00.000Z"
        updatedAt:
          type: string
          format: date-time
          description: Last update timestamp (ISO 8601)
          example: "2024-01-15T10:30:00.000Z"
      required:
        - id
        - email
        - name
        - role
        - createdAt
        - updatedAt
...
    UpdateUser:
      type: object
      properties:
        email:
          type: string
          format: email
          maxLength: 255
          description: User email address
          example: "john.updated@example.com"
        name:
          type: string
          minLength: 1
          maxLength: 255
          description: User's full name
          example: "John Updated"
        role:
          type: string
          enum: [customer, admin]
          description: User role
          example: "admin"
...
    Error:
      type: object
      properties:
        error:
          type: string
          description: Error message
          example: "Resource not found"
      required:
        - error